Update to release v.3

+++ run_wordpress.sh
@@ -2,34 +2,22 @@
 
 # Vars
 FILE=/var/www/html/wp-content
-DB=/var/lib/mysql/mysql/db.frm
-DBWP=/var/lib/mysql/wordpress/wp_users.frm
+DBWP=/var/lib/mysql/wordpress/wp_users.idb
 DIR_BCK=/var/www/backup
 WP_INST=false
 WP_RESTORE=false
 
 
 # Remove crashed container run pid
-rm /var/run/apache2/apache2.pid
-
-
-
-# Check DB then copy db files
-if [ -a $DB ];
-then
-	echo "DB exists"
-else
-	echo "Copy DB"
-	cp -a /var/lib/mysql-old/* /var/lib/mysql/
-fi
+[ -f /var/run/apache2/apache2.pid ] && rm /var/run/apache2/apache2.pid
 
 
 # Check wp is installed
-if [ -a $DBWP ];
+if [ -f $DBWP ];
 then
-	echo "DB wp exists"
+	echo "Wordpress already installed"
 else
-	echo "Must install wp"
+	echo "Proceed with installation or restore"
 	WP_INST=true
 fi
 
@@ -49,16 +37,21 @@ else
 
 	# Check for backup
 	if [ "$(ls -A $DIR_BCK)" ]; then
-		echo "Try restoring backup"
+		echo "Try restoring backup... "
 		cp -a /var/www/backup/* /var/www/html/
 		chown www-data /var/www/html/.
 		WP_RESTORE=true
 		WP_INST=false
-	else
-		echo "Installing wp"
+	elif [ ! -f /var/www/html/wp-config.php ]; then
+		echo "Installing Wordpress"
    		#cp -a /var/www/html-wp/* /var/www/html/
-   		rm  /var/www/html/index.html
+   		[ -f /var/www/html/index.html ] && rm  /var/www/html/index.html
    		chown www-data /var/www/html/.
+		WP_RESTORE=false
+		WP_INST=true
+	else
+	    WP_RESTORE=false
+		WP_INST=false
 	fi
    	
 fi
@@ -67,6 +60,11 @@ if [ $WP_RESTORE == "true" ];
 then
 	echo "Executing restore create database"
 	mysql -u $DB_ROOT_USER -e "create database $DB_NAME;"
+
+	mysql -u $DB_ROOT_USER -e "create user 'wordpress'@'127.0.0.1' identified by '$DB_ROOT_PWD';"
+	mysql -u $DB_ROOT_USER -e "grant all privileges on $DB_NAME.* to 'wordpress'@'127.0.0.1';"
+	mysql -u $DB_ROOT_USER -e "flush privileges;"
+	
 	echo "Run installer via web"
 elif [ $WP_INST == "true" ];
 then
@@ -74,8 +72,12 @@ then
 	#sudo -u www-data -s -- /usr/local/bin/wp core download --locale="$WP_LOCALE"
    	sudo cp -a /var/www/html-wp/* /var/www/html/
    	mysql -u $DB_ROOT_USER -e "create database $DB_NAME;"
+	
+	mysql -u $DB_ROOT_USER -e "create user 'wordpress'@'127.0.0.1' identified by '$DB_ROOT_PWD';"
+	mysql -u $DB_ROOT_USER -e "grant all privileges on $DB_NAME.* to 'wordpress'@'127.0.0.1';"
+	mysql -u $DB_ROOT_USER -e "flush privileges;"
    
-   	sudo -u www-data -s -- /usr/local/bin/wp core config --dbname=$DB_NAME --dbuser=$DB_ROOT_USER
+   	sudo -u www-data -s -- /usr/local/bin/wp core config --dbname=$DB_NAME --dbuser=wordpress --dbpass=$DB_ROOT_PWD --dbhost=127.0.0.1
    	
    	sudo -u www-data -s -- /usr/local/bin/wp core install \
    		--url=http://$WP_IPFDQN/ \
@@ -83,20 +85,10 @@ then
    		--admin_user=$WP_ADMIN_USER \
    		--admin_password=$WP_ADMIN_PWD \
    		--admin_email=$WP_ADMIN_EMAIL 
-
 fi 
 
-if [ -a /var/www/html/index.html ]; then
-	rm  /var/www/html/index.html
-fi
+[ -f /var/www/html/index.html ] && rm  /var/www/html/index.html
 
-if [ $WP_RESTORE == "false" ]; 
-then
-	if [ -a /var/www/html/installer.php ]; 
-	then
-		rm /var/www/html/*.zip
-	fi
-fi
 
 # Apache 
 source /etc/apache2/envvars
